poprawki w utils: poprawny typ wartości dla czestotliwosci probkowania kostiumu poprawka dla statusu próbek - wyciąganie statusu przy braku lower_bound poprawki dla strumienie - push danych z move semtics zniesienie wymogu dla operatorów strumieni do bycia const usuniecie zależności od boost circular buffer - zamiana na list, boost alokował z góry zadaną ilośc elementów
poprawki dla imu:
nagrywanie surowych danych do własnego formatu
poprawki dla aktualizacji stanu połaczenia
poprawki dla przetwarzania danych i przekazywania danych przez strumieenie

git-svn-id: https://localhost:8443/svn/hm-mde-1@1245 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ branches/kinematicRefactorForIMU/src/edrutils/src/imucostumelib/src/CostumeRawIO.cpp
@@ -131,7 +131,7 @@ public:
 	//OK
 	void setSamplingDelay(const uint32_t delay, const uint16_t timeout)
 	{
-		auto d = utils::clamp<uint32_t>(delay / 10, 1, 255);
+		auto d = utils::clamp<uint8_t>(delay / 10, 1, 255);
 		std::array<uint8_t, 19> buffer;
 		buffer[0] = 0x04;
 		buffer[1] = (uint8_t)this;
